replaced streamlit solver with MCKP solver
