+++ streamlit.py
@@ -1,13 +1,12 @@
 import numpy as np
 import pandas as pd
 import streamlit as st
-from ortools.linear_solver import pywraplp
-from utils.linear_programming import round_decimals_up, linearoptimizer
+
+from math import ceil
 from utils.utils import calculate_alternative, survey_extension_solver
 
-
 st.write(
-    '### CALL ANALYST OPTIMIZATION MODEL (CAP MODEL)\n'
+    '### CALL ANALYST ALLOCATION OPTMIZATION\n'
     '\n'
     '#### Description\n'
     'An optimization model that will guide our call analyst '
@@ -15,235 +14,240 @@ st.write(
     'survey projects in a timely manner.\n'
     '***'
 )
+col1, col2 = st.columns(2)
 
-st.sidebar.write('''## STEP 1: Input parameters''')
-Number_running_survey = st.sidebar.number_input(
-    'How many surveys will be conducted today? (Max: 10)',
-    1,
-    10
-)
-ManPower = st.sidebar.number_input(
-    'How many call agents we have? (Max: 100)',
-    1,
-    100
-)
-today = st.sidebar.date_input(
-    'Today\'s date',
-    key='current_date'
-)
-st.sidebar.write('''
-***
-## STEP 2: Survey input details
-''')
+def survey_input_details() -> pd.DataFrame:
+    survey_names = []
+    cr_rates = []
+    remaining_cr = []
+    remaining_working_days = []
+    target_cr_day = []
+    allocated_agents = []
 
-def Survey_input_details():
-    Names = []
-    CRs = []
-    NoCallReqs = []
-    NoDaysRemains = []
-    MinCalls = []
-    ManpowerAllocateds = []
 
-    for i in range(Number_running_survey):
-        st.sidebar.write('#### Please enter survey details as below:')
-
-        Name = st.sidebar.text_input(f'{i}. Survey name ')
-        CR = st.sidebar.slider(
-            f'{i}. Avg Daily CR/agent ',
-            1,
-            50,
-            10
-        )
-        NoCallReq = st.sidebar.number_input(
-            f'{i}. Remaining CR ',
-            1,
-            10000
-        )
-        DueDate = st.sidebar.date_input(
-            f'{i}. Target Completion Date', 
-            value=np.busday_offset(today, 1, 'forward').tolist(), 
-            min_value=today, 
-            key=f'{i}_complete_date'
+    with col1:
+        extension = st.number_input(
+            'Maximum Survey Duration',
+            min_value=0,
+            max_value=180,
+            value=10,
+            key='max_extension'
         )
 
-        PlannedCall = int()
-        ManpowerAllocated = int()
-        Names.append(Name)
-        CRs.append(CR) 
-        NoCallReqs.append(int(NoCallReq))
-        NoDaysRemain = np.busday_count(today, DueDate)
-        NoDaysRemains.append(NoDaysRemain)
+    with st.sidebar:
+        st.write('## STEP 1: Input parameters')
+        input_type = st.selectbox(
+            label='CSV Upload or Manual Input',
+            options=['CSV', 'Manual'],
+            index=1,
+            key='input_type'
+        )
 
-        if NoDaysRemain != 0:
-            MinCall = round_decimals_up(NoCallReq / NoDaysRemain)
-            MinCalls.append(int(MinCall))
-        else:
-            MinCall = NoCallReq / NoDaysRemain
-            MinCalls.append(MinCall)
+        if input_type == 'CSV':
+            survey_csv = st.file_uploader(
+                label='Upload survey CSV file',
+                type='csv',
+                key='csv_file'
+            )
+            manpower = st.number_input(
+                'How many call agents we have? (Max: 100)',
+                1,
+                100,
+                key='manpower'
+            )
 
-        ManpowerAllocateds.append(ManpowerAllocated)
-        st.sidebar.write('***')
+        if st.session_state.get('csv_file', False):
+            survey_df = pd.read_csv(survey_csv, parse_dates=[4], nrows=10)
+            survey_df.set_index('index', inplace=True)
+            survey_df.dropna(inplace=True)
+
+            today = st.date_input(
+                'Today\'s date',
+                disabled=True,
+                key='current_date'
+            )
+
+            survey_names = survey_df['survey_title'].tolist()
+            cr_rates = survey_df['cr_rate_per_agent'].tolist()
+            remaining_cr = survey_df['remaining_cr'].astype(int).tolist()
+            survey_day_count = np.busday_count(
+                today,
+                survey_df['deadline'].values.astype('<M8[D]')
+            )
+            remaining_working_days = survey_day_count.tolist()
+
+            survey_target_cr = np.ceil(survey_df['remaining_cr'] / survey_day_count)
+            target_cr_day = survey_target_cr.values.astype(int).tolist()
+            allocated_agents = [int() for _ in survey_names]
+            survey_planned_cr = int()
+
+        if input_type == 'Manual':
+            survey_count = st.number_input(
+                'How many surveys will be conducted today? (Max: 10)',
+                1,
+                10,
+                key='survey_count'
+            )
+            manpower = st.number_input(
+                'How many call agents we have? (Max: 100)',
+                1,
+                100,
+                key='manpower'
+            )
+            today = st.date_input(
+                'Today\'s date',
+                key='current_date'
+            )
+            st.write(
+            '''
+            ***
+            ## STEP 2: Survey input details
+            '''
+            )
+
+            for i in range(survey_count):
+                st.write('#### Please enter survey details as below:')
+
+                name = st.text_input(f'{i}. Survey name ')
+                survey_cr_rate = st.number_input(
+                    f'{i}. Avg Daily CR/agent ',
+                    1,
+                    50,
+                    10
+                )
+                survey_cr = st.number_input(
+                    f'{i}. Remaining CR ',
+                    1,
+                    10000
+                )
+                # deadline = st.checkbox(
+                #     'Hard Deadline?',
+                #     value=False,
+                #     key=f"deadline_{i}"
+                # )
+                due_date = st.date_input(
+                    f'{i}. Target Completion Date', 
+                    value=np.busday_offset(today, 1, 'forward').tolist(), 
+                    min_value=today, 
+                    key=f'{i}_complete_date',
+                    # disabled=deadline
+                )
+
+                survey_planned_cr = int()
+                survey_agents = int()
+                survey_names.append(name)
+                cr_rates.append(survey_cr_rate) 
+                remaining_cr.append(int(survey_cr))
+                survey_day_count = np.busday_count(today, due_date)
+                remaining_working_days.append(survey_day_count)
+
+                if survey_day_count != 0:
+                    survey_target_cr = ceil(survey_cr / survey_day_count)
+                    target_cr_day.append(int(survey_target_cr))
+                else:
+                    survey_target_cr = survey_cr / survey_day_count
+                    target_cr_day.append(survey_target_cr)
+                allocated_agents.append(survey_agents)
+                st.write('***')
 
     data = {
-        'Survey Title': Names,
-        'Avg Daily CR/agent': CRs,
-        'Remaining CR': NoCallReqs,
-        'Remaining Working Days': NoDaysRemains,
-        'Target CR/day': MinCalls,
-        'Call Agents Allocation': ManpowerAllocateds,
-        'Plan CR/day': PlannedCall
+        'Survey Title': survey_names,
+        'Avg Daily CR/agent': cr_rates,
+        'Remaining CR': remaining_cr,
+        'Remaining Working Days': remaining_working_days,
+        'Call Agents Allocation': allocated_agents,
+        'Target CR/day': target_cr_day,
+        'Plan CR/day': survey_planned_cr,
+        'Deadline Days': survey_day_count
     }
 
     dataframe = pd.DataFrame.from_dict(data)
     return dataframe
 
 
-def View_results():
-    if df.iloc[0, 0] == '':
-        st.write('\nPlease fill in the required information\n')
-        return None
+def display_solution(df: pd.DataFrame):
+    remaining_cr = df['Remaining CR'].tolist()
+    cr_rates = df['Avg Daily CR/agent'].tolist()
+    starting_day = [0 for _ in range(df.shape[0])]
 
-    if results == pywraplp.Solver.OPTIMAL:
-        st.write('#### Surveys Summary')
-
-        for y in range(Number_running_survey):
-            df.iloc[y,5] = SolutionValues[y]
-            df.iloc[y,6] = SolutionValues[y] * df.iloc[y,1]
-
-        st.write(
-            'The algorithm found the optimal call analysts allocation'
-            ' to maximize the total plan CR/day & meet the surveys target'
-            ' completion date.'
-        )
-        st.table(df)
-        st.write('##### Target vs Plan CR/day')
-        columns = [*st.columns(Number_running_survey)]
-
-        for i in range(Number_running_survey):
-            try:
-                columns[i].write(df.iloc[i, 0])
-                columns[i].metric(
-                    label='Target CR/day',
-                    value=df.iloc[i, 4],
-                    help=None
-                )
-                columns[i].metric(
-                    label='Plan CR/day',
-                    value=df.iloc[i, 6],
-                    delta=int(df.iloc[i, 6] - df.iloc[i, 4]),
-                    delta_color='normal',
-                    help=None
-                )
-
-                agent_difference = (df.iloc[i, 6] - df.iloc[i, 4]) // df.iloc[i, 1] or "0"
-                agent_color = 'off' if agent_difference in ('-', '0', 'None') else 'normal'
-                columns[i].metric(
-                    label='Agents',
-                    value=df.iloc[i, 5],
-                    delta=str(agent_difference),
-                    delta_color=agent_color,
-                    help=None
-                )
-
-            except Exception as e:
-                raise e
-                pass
-
-        st.write('\n')
-        st.write('##### Remark')
-        st.write(
-            'The optimization only to make sure all surveys can be'
-            ' completed exactly on the target completion date with'
-            ' maximum number of total plan CR / day.\nIf the'
-            ' different between **plan CR / day** and **target CR / day**'
-            ' are skewed toward any of surveys, you can make adjustment'
-            ' by tuning the **Target Completion Date** of each surveys.'
-        )
-
-    else:
-        st.write('##### Surveys Summary')
-        df.iloc[:,[5, 6]] = 'Infeasible'
-
-        st.write('\n')
-        st.write(
-            'The algorithm terminated successfully and determined '
-            'that the problem is infeasible.'
-        )
-        st.table(df)
-        st.write('\n')
-        st.write('##### Remark')
-        st.write(
-            'The algorithm unable to find the optimal call agents '
-            'allocation to meet the surveys dateline.\nTo re-evaluate '
-            'the opmitization, You may replan by:\n1. extending the '
-            'surveys **target completion date**.\n2. Increasing the '
-            '**total number of call agents**.'
-        )
-
-
-def View_Optimal():
-    survey_cr = df['Remaining CR'].tolist()
-    survey_rate = df['Avg Daily CR/agent'].tolist()
-    survey_day = [0 for _ in range(df.shape[0])]
-    
-    surveys = list(zip(survey_cr, survey_rate, survey_day))
+    surveys = list(zip(remaining_cr, cr_rates, starting_day))
     max_extension = st.session_state.get('max_extension', 10)
-    max_manpower = ManPower
+    max_manpower = st.session_state.get('manpower', 1)
 
     solutions = survey_extension_solver(
         surveys=surveys,
         max_extension=max_extension,
         max_manpower=max_manpower,
+        top_n=5
     )
+
+
     if solutions:
-        solution_df = pd.DataFrame()
-        agents, days = solutions[0]
-        solution_df = solution_df.assign(
-            **{
+        solution_options = [f'Top {i+1}' for i in range(min(5, len(solutions)))]
+
+        with col2:
+            st.selectbox(
+                'Select top solutions',
+                options=solution_options,
+                index=0,
+                key='top_n'
+            )
+
+        top_n = solution_options.index(st.session_state.get('top_n', 'Top 1'))
+        agents, days = solutions[top_n]
+        today = st.session_state.get('current_date', None)
+        due_dates = np.busday_offset(
+            dates=today,
+            offsets=days,
+            roll='forward'
+        )
+        leftover_cr = (df['Deadline Days'] * df['Avg Daily CR/agent'] * pd.Series(agents))
+        solutions_df = pd.DataFrame.from_dict(
+            {
                 'Survey Title': df['Survey Title'],
                 'Avg Daily CR/agent': df['Avg Daily CR/agent'],
                 'Remaining CR': df['Remaining CR'],
-                'End Date': np.busday_offset(dates=today, offsets=days, roll='forward'),
-                'Target CR/day': df['Remaining CR'] // pd.Series(days),
+                'Total CR by Deadline': leftover_cr,
+                'Remaining CR by Deadline': df['Remaining CR'] - leftover_cr,
+                'Remaining Working Days': pd.Series(days),
+                'Optimal Due Date': due_dates,
                 'Call Agents Allocation': agents,
-                'Plan CR/day': df['Avg Daily CR/agent'] * pd.Series(agents)
-            },
+                'Target CR/day': np.ceil(df['Remaining CR'] / df['Deadline Days']),
+                'Target CR/Agent/day': np.ceil(df['Remaining CR'] / df['Deadline Days'] / agents),
+                'Plan CR/day': df['Avg Daily CR/agent'] * pd.Series(agents),
+            }
         )
-        solution_df = solution_df.astype(
+
+        solutions_df = solutions_df.astype(
             {
                 'Survey Title':'string',
                 'Avg Daily CR/agent': 'int',
                 'Remaining CR': 'int',
-                'End Date': 'string',
-                'Target CR/day': 'int',
+                'Total CR by Deadline': 'int',
+                'Remaining CR by Deadline': 'int',
+                'Remaining Working Days': 'int',
+                'Optimal Due Date': 'string',
                 'Call Agents Allocation': 'int',
+                'Target CR/day': 'int',
+                'Target CR/Agent/day': 'int',
                 'Plan CR/day': 'int',
             },
         )
-        st.table(solution_df)
+        st.table(solutions_df)
     else:
-        st.write("Solution not found. Either extend the maximum extension or increase manpower.")
+        st.write(
+            'Solution not found.'
+            ' Either extend the maximum extension or increase manpower.'
+        )
 
-# Get Input from user
+
+# Main display
 try:
-    df = Survey_input_details()
-except Exception as e1:
-    st.error(f'e1: {e1}') 
-
-# Calculate Mixed Integer Linear Programming
+    df = survey_input_details()
+except UnboundLocalError:
+    pass
 try:
-    results, SolutionValues = linearoptimizer(df, Number_running_survey, ManPower)
-except Exception as e2:
-    st.error(f'e2: {e2}')
-
-
-# View result
-try:
-    veiw = View_results()
-    if df.iloc[0, 5] == 'Infeasible' and (df.iloc[:, 3] > 0).all():
-        calculate_alternative(df, ManPower, today)
-        View_Optimal()
-
-except Exception as e3:
-    raise e3
-    st.error(f'e3: {e3}')
+    if df['Survey Title'].all():
+        display_solution(df)
+except NameError:
+    pass
